* implement remaining GTK_NOT_IMPLEMENTED and GTK_PARTIALLY_IMPLEMENTED    - AW_root::window_show() and window_hide()    - AW_at_max_size (restore from trunk)    - AW_at_auto (restore from trunk)    - AW_window::clear_option_menu()    - AW_window::highlight();    - AW_window::local_id();    - AW_AREA_BOTTOM for PHYLO    - AW_window::set_focus_callback    - AW_window::draw_line()      * needed for user masks only      * fix AW_window::get_window_size -- it's an AW_at function really      * fix AW_window::load_xfig(NULL) (scaling)    - AW_window::set_bottom_area_size      * needed a work-around because PH_phylo calls it with        a fixed height not accounting for font_height!=13    - AW_window recalc_pos_at_show    - AW_advice.cxx restored from trunk    - AW_question.cxx restored from trunk    - AW_edit.cxx restored from trunk    - AW_root_debug.cxx restored from trunk (AW_debug.cxx)    - AW_window_debug.cxx removed (mnemonics can be duplicate in GTK)    - AW_root::save_properties restored from trunk    - AW_preset::* restore from trunk * fix warnings * remove toggle focus from NTREE * use gtk file chooser in aw_modal (store/save config from AWT) * FIXME: unit tests seem to be broken now

+++ WINDOW/AW_motif_gtk_conversion.cxx
@@ -11,7 +11,7 @@ std::string AW_motif_gtk_conversion::convert_mnemonic(const char* text, const ch
     std::string strText(text);
     std::string strMnemonic(mnemonic);
     if(NULL != mnemonic) {
-        int mnemonicPos = strText.find(strMnemonic);
+        size_t mnemonicPos = strText.find(strMnemonic);
         if (strText.npos == mnemonicPos) {
             //mnemonic string is not part of the name, append it
             strText = strText + "(_" + strMnemonic + ")";
